backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1
google-genai==1.1.0
aiosqlite==0.20.0
python-docx==1.1.2
PyPDF2==3.0.1
httpx==0.28.1
